Start from scratch
This banch aims at wrapping the library using ctypes instead of
cython. The reason for this change is dependency management,
specifically cython requires that the C library be installed
and in the library path before pyinform can be built. Further, you
have to have a C compiler to use cython. This might not be a
difficult dependency to satisfy, but if we are going to target
Windows (and OS X), then there is no promise that the system has
a C compiler and that it is easy to access.

With ctypes we can (in principle) have setup.py pull down the
binary dependencies for us. Problems solved! Hopefully.