Ensure that arrays are in C-contiguous form
Sometimes NumPy arrays change memory layout when they are
manipulated, e.g. when they are transposed. This can result in
insidious bugs when you pass those arrays to C functions. To
ensure this does not happen, we force the arrays into a
C-contiguous format before we start to operate on them. This
also ensures that we don't have to call `ravel` to bring them
down to 1D.