Fix access errors
The Inform v1.0.0 API for conditional entropy and relative entropy
changed from v0.0.5. In the previous version, the base of the
logarithm was passed as an argument. Now, all logarithms are
computed base-2. That change was not reflected in the ctypes call
signatures nor in the associated calls in the python functions.